+++ xclim/core/calendar.py
@@ -6,9 +6,8 @@ Calendar handling utilities
 
 Helper function to handle dates, times and different calendars with xarray.
 """
-
 import datetime as pydt
-from typing import Optional, Sequence, Union
+from typing import Any, Optional, Sequence, Union
 from warnings import warn
 
 import cftime
@@ -28,17 +27,7 @@ from xarray.coding.cftimeindex import CFTimeIndex
 from xarray.core.resample import DataArrayResample
 
 # cftime and datetime classes to use for each calendar name
-datetime_classes = {
-    "default": pydt.datetime,
-    "standard": cftime.DatetimeGregorian,
-    "gregorian": cftime.DatetimeGregorian,
-    "proleptic_gregorian": cftime.DatetimeProlepticGregorian,
-    "julian": cftime.DatetimeJulian,
-    "noleap": cftime.DatetimeNoLeap,
-    "all_leap": cftime.DatetimeAllLeap,
-    "360_day": cftime.Datetime360Day,
-}
-
+datetime_classes = {"default": pydt.datetime, **cftime._cftime.DATE_TYPES}
 
 # Maximum day of year in each calendar.
 max_doy = {
@@ -48,56 +37,69 @@ max_doy = {
     "proleptic_gregorian": 366,
     "julian": 366,
     "noleap": 365,
+    "365_day": 365,
     "all_leap": 366,
+    "366_day": 366,
     "360_day": 360,
 }
 
 
-def get_calendar(arr: Union[xr.DataArray, xr.Dataset], dim: str = "time") -> str:
-    """Return the calendar of the time coord of the DataArray.
+def get_calendar(obj: Any, dim: str = "time") -> str:
+    """Return the calendar of an object.
 
     Parameters
     ----------
-    arr : Union[xr.DataArray, xr.Dataset]
-      Array/dataset with a datetime coordinate. Values must either be of datetime64 dtype or have a dt.calendar attribute.
+    obj : Any
+      An object defining some date..
+      If `obj` is an array/dataset with a datetime coordinate, use `dim` to specify its name.
+        Values must have either a datetime64 dtype or a cftime dtype.
+      `obj`can also be a python datetime.datetime, a cftime object or a pandas Timestamp
+        or an iterable of those, in which case the calendar is inferred from the first value.
     dim : str
-      Name of the coordinate to check.
+      Name of the coordinate to check (if `obj` is a DataArray or Dataset).
 
     Raises
     ------
     ValueError
-        If `arr` doesn't have a datetime64 or cftime dtype.
+      If no calendar could be inferred.
 
     Returns
     -------
     str
-      The cftime calendar name or "default" when the data is using numpy's datetime type (numpy.datetime64.
+      The cftime calendar name or "default" when the data is using numpy's or python's datetime types.
     """
-    if arr[dim].dtype == "O":  # Assume cftime, if it fails, not our fault
-        non_na_item = arr[dim].where(arr[dim].notnull(), drop=True)[0].item()
-        cal = non_na_item.calendar
-    elif "datetime64" in arr[dim].dtype.name:
-        cal = "default"
-    else:
-        raise ValueError(
-            f"Cannot infer calendars from timeseries of type {arr[dim][0].dtype}"
-        )
-    return cal
+    if isinstance(obj, (xr.DataArray, xr.Dataset)):
+        if obj[dim].dtype == "O":
+            obj = obj[dim].where(obj[dim].notnull(), drop=True)[0].item()
+        elif "datetime64" in obj[dim].dtype.name:
+            return "default"
+
+    obj = np.take(
+        obj, 0
+    )  # Take zeroth element, overcome cases when arrays or lists are passed.
+    if isinstance(obj, pydt.datetime):  # Also covers pandas Timestamp
+        return "default"
+    if isinstance(obj, cftime.datetime):
+        return obj.calendar
+
+    raise ValueError(f"Calendar could not be inferred from object of type {type(obj)}.")
 
 
 def convert_calendar(
     source: Union[xr.DataArray, xr.Dataset],
     target: Union[xr.DataArray, str],
     align_on: Optional[str] = None,
+    missing: Optional[Any] = None,
     dim: str = "time",
 ) -> xr.DataArray:
     """Convert a DataArray/Dataset to another calendar using the specified method.
 
-    Only converts the individual timestamps, does not modify any data except in dropping invalid/surplus dates.
+    Only converts the individual timestamps, does not modify any data except in dropping invalid/surplus dates or inserting missing dates.
 
     If the source and target calendars are either no_leap, all_leap or a standard type, only the type of the time array is modified.
     When converting to a leap year from a non-leap year, the 29th of February is removed from the array.
-    In the other direction and if `target` is a string, the 29th of February will be missing in the output.
+    In the other direction and if `target` is a string, the 29th of February will be missing in the output,
+    unless `missing` is specified, in which case that value is inserted.
 
     For conversions involving `360_day` calendars, see Notes.
 
@@ -106,13 +108,16 @@ def convert_calendar(
     Parameters
     ----------
     source : xr.DataArray
-      Input array/dataset with a time coordinate of a valid dtype (datetime64 or a cftime.datetime)
+      Input array/dataset with a time coordinate of a valid dtype (datetime64 or a cftime.datetime).
     target : Union[xr.DataArray, str]
       Either a calendar name or the 1D time coordinate to convert to.
       If an array is provided, the output will be reindexed using it and in that case, days in `target`
-      that are missing in the converted `source` are filled by NaNs.
+      that are missing in the converted `source` are filled by `missing` (which defaults to NaN).
     align_on : {None, 'date', 'year'}
       Must be specified when either source or target is a `360_day` calendar, ignored otherwise. See Notes.
+    missing : Optional[any]
+      A value to use for filling in dates in the target that were missing in the source.
+      If `target` is a string, default (None) is not to fill values. If it is an array, default is to fill with NaN.
     dim : str
       Name of the time coordinate.
 
@@ -120,9 +125,10 @@ def convert_calendar(
     -------
     Union[xr.DataArray, xr.Dataset]
       Copy of source with the time coordinate converted to the target calendar.
-      The length of the array is the same as `target` if an array was given, otherwise it stays the same as `source`.
-      Except if source is a `360_day` calendar and `align_on='date'`: then a daily source will be output with 358 dates
-      per year on a non leap year, 359 on a leap year, see Notes.
+      If `target` is given as an array, the output is reindexed to it, with fill value `missing`.
+      If `target` was a string and `missing` was None (default), invalid dates in the new calendar are dropped, but missing dates are not inserted.
+      If `target` was a string and `missing` was given, then start, end and frequency of the new time axis are inferred and
+         the output is reindexed to that a new array.
 
     Notes
     -----
@@ -164,19 +170,15 @@ def convert_calendar(
     if cal_src == cal_tgt:
         return source
 
-    out = source.copy()
     if (cal_src == "360_day" or cal_tgt == "360_day") and align_on is None:
         raise ValueError(
             "Argument `align_on` must be specified with either 'date' or "
             "'year' when converting to or from a '360_day' calendar."
         )
-    if (cal_src != "360_day" and cal_tgt != "360_day") and align_on is not None:
-        warn(
-            "Argument `align_on` was specified, but none of the source or "
-            "target calendars is '360_day'. `align_on` will be ignored."
-        )
+    if cal_src != "360_day" and cal_tgt != "360_day":
         align_on = None
 
+    out = source.copy()
     # TODO Maybe the 5-6 days to remove could be given by the user?
     if align_on == "year":
 
@@ -212,8 +214,15 @@ def convert_calendar(
         # Remove NaN that where put on invalid dates in target calendar
         out = out.where(out[dim].notnull(), drop=True)
 
+    if isinstance(target, str) and missing is not None:
+        target = date_range_like(source[dim], cal_tgt)
+
     if isinstance(target, xr.DataArray):
-        out = out.reindex({dim: target})
+        out = out.reindex({dim: target}, fill_value=missing or np.nan)
+
+    # Copy attrs but change remove `calendar` is still present.
+    out[dim].attrs.update(source[dim].attrs)
+    out[dim].attrs.pop("calendar", None)
     return out
 
 
@@ -254,6 +263,72 @@ def interp_calendar(
     out = out.interp(time=target_idx)
     out[dim] = target
     return out
+
+
+def date_range(*args, calendar="default", **kwargs):
+    """Wrap pd.date_range (if calendar == 'default') or xr.cftime_range (otherwise)."""
+    if calendar == "default":
+        return pd.date_range(*args, **kwargs)
+    return xr.cftime_range(*args, calendar=calendar, **kwargs)
+
+
+def date_range_like(source, calendar):
+    """Generate a datetime array with the same frequency, start and end as another one, but in a different calendar.
+
+    Parameters
+    ----------
+    source : xr.DataArray
+      1D datetime coordinate DataArray
+    calendar : str
+      New calendar name.
+
+    Raises
+    ------
+    ValueError
+      If the source's frequency was not found.
+
+    Returns
+    -------
+    xr.DataArray
+      1D datetime coordinate with the same start, end and frequency as the source, but in the new calendar.
+        The start date is assumed to exist in the target calendar.
+        If the end date doesn't exist, the code tries 1 and 2 calendar days before.
+        Exception when the source is in 360_day and the end of the range is the 30th of a 31-days month,
+        then the 31st is appended to the range.
+    """
+    freq = xr.infer_freq(source)
+    if freq is None:
+        raise ValueError(
+            "`date_range_like` was unable to generate a range as the source frequency was not inferrable."
+        )
+
+    src_cal = get_calendar(source)
+    if src_cal == calendar:
+        return source
+
+    index = source.indexes[source.dims[0]]
+    end_src = index[-1]
+    end = _convert_datetime(end_src, calendar=calendar)
+    if end is np.nan:  # Day is invalid, happens at the end of months.
+        end = _convert_datetime(end_src.replace(day=end_src.day - 1), calendar=calendar)
+        if end is np.nan:  # Still invalid : 360_day to non-leap february.
+            end = _convert_datetime(
+                end_src.replace(day=end_src.day - 2), calendar=calendar
+            )
+    if src_cal == "360_day" and end_src.day == 30 and end.daysinmonth == 31:
+        # For the specific case of daily data from 360_day source, the last day is expected to be "missing"
+        end = end.replace(day=31)
+
+    return xr.DataArray(
+        date_range(
+            _convert_datetime(index[0], calendar=calendar),
+            end,
+            freq=freq,
+            calendar=calendar,
+        ),
+        dims=source.dims,
+        name=source.dims[0],
+    )
 
 
 def _convert_datetime(